Nhom8_MayKafe_WebServer/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
django-cors-headers>=4.4,<4.5
